Rename 'database.tables' to 'database.models'
